jubilant/statement_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
supabase==2.6.0
pdfplumber==0.11.4
camelot-py==0.11.0
opencv-python==4.10.0.84
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
python-dateutil==2.9.0.post0
